refactor(extensions): extract the skills modal farm to its own component
ExtensionsPanel rendered 17 GlassModals inline; the five Skills-tab
modals (new-skill scaffold, editor, save feedback, discard confirm,
write-conflict confirm — ~440 lines of JSX) moved verbatim into
ExtensionsPanelSkillModals with one typed props bundle for the skill
state cluster. The three extensions-surface guard tests keep passing
unchanged: the tab surface and the plugins/MCP modals stay in the
panel.

ExtensionsPanel: 4,391 → 3,980 lines. Part of the giant-component
decomposition series; the plugins/MCP modal farm follows.
